finishing part4

+++ part4/bloglist-backend/controllers/blogs.js
@@ -2,7 +2,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   res.json(blogs)
 })
 
@@ -14,33 +14,60 @@ blogsRouter.get('/:id', async (req, res) => {
 })
 
 blogsRouter.post('/', async (req, res) => {
-  const body = req.body
+  const user = req.user
+  if (!user) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
 
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes
+    title: req.body.title,
+    author: req.body.author,
+    url: req.body.url,
+    likes: req.body.likes || 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   res.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
-  const blogToDelete = await Blog.findByIdAndDelete(req.params.id)
-  !blogToDelete
-    ? res.status(404).end()
-    : res.status(204).end()
+  const user = req.user
+  if (!user) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(req.params.id)
+  if (!blog) {
+    return res.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return res.status(403).json({ error: 'unauthorized: not the blog owner' })
+  }
+
+  await Blog.findByIdAndDelete(req.params.id)
+  res.status(204).end()
 })
 
 blogsRouter.put('/:id', async (req, res) => {
-  const { likes } = req.body
+  const user = req.user
+  if (!user) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
 
+  const { likes } = req.body
   const blog = await Blog.findByIdAndUpdate(req.params.id)
 
   if (!blog) {
-    return res.status(404).end()
+    return res.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return res.status(403).json({ error: 'unauthorized: not the blog owner' })
   }
 
   blog.likes = likes
